Fixed crash from div by 0 when window was minimized
On Windows, when a window is minimized, GLFW reports a window size `w, h = (0, 0)`. Here, we're dividing by w and h, so upon minimization the app crashed from a division by zero.

NOTE: I'm not sure what `io.display_fb_scale` should be when the window is minimized, but since nothing is rendered, maybe we can leave it as it was before minimizing?