Catch when modifiers are pressed in pyglet
If we want to catch a modifier being pressed, for example via
io.key_alt, we need to catch when the alt key is pressed by itself.
Pyglet fires events for the key press of alt, but in those cases it's
not flipped on in the modifiers argument.  (Conversely, it *is* on when
key release is fired when you let go of it.)

This catches that and sets the io attribute appropriately.